LPS-96228 Move FragmentEntry into its own TestUtil
